removed the main.cpp file, and seperated the pathfinder interface from the a* implementation
